requirements.txt
# Microsoft Copilot Studio Client SDK - Python Dependencies
# Based on official Microsoft Agents SDK sample
# Python 3.9+ Required

# Copilot Studio Client (for invoking your agent directly)
microsoft-agents-copilotstudio-client
microsoft-agents-activity

# Authentication
msal>=1.20.0
azure-identity>=1.18.0

# HTTP client
aiohttp>=3.8.0

# Configuration management
PyYAML>=6.0.2
python-dotenv>=1.0.1

# Type hints support
typing-extensions>=4.12.0; python_version < "3.11"

# FastAPI (for REST API server)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0

# Development dependencies (optional)
# pytest>=8.3.0
# pytest-asyncio>=0.24.0
# black>=24.8.0
# mypy>=1.11.0
